+++ tests/delete.rs
@@ -1,5 +1,6 @@
+use similar_asserts::assert_eq;
 use std::error::Error;
-use wayfind::Router;
+use wayfind::{errors::DeleteError, Router};
 
 #[test]
 fn expanded() -> Result<(), Box<dyn Error>> {
@@ -16,22 +17,29 @@ fn expanded() -> Result<(), Box<dyn Error>> {
                      ╰─ / ○
     "#);
 
-    // Should not be able to delete via expande routes.
-    router.delete("/files/{name}/")?;
+    assert_eq!(
+        router.delete("/files/{name}/").unwrap_err(),
+        DeleteError::RouteMismatch {
+            route: "/files/{name}/".to_string(),
+            inserted: "/files/{name}.{extension?}{/}".to_string(),
+        }
+    );
 
     insta::assert_snapshot!(router, @r#"
     ▽
     ╰─ /files/
              ╰─ {name} ○
-                     ╰─ .
-                        ╰─ {extension} ○
-                                     ╰─ / ○
+                     ├─ .
+                     │  ╰─ {extension} ○
+                     │               ╰─ / ○
+                     ╰─ / ○
     "#);
 
-    // Should be able to delete via pre-expanded route.
     router.delete("/files/{name}.{extension?}{/}")?;
 
-    insta::assert_snapshot!(router, @"");
+    insta::assert_snapshot!(router, @"
+    ▽
+    ");
 
     Ok(())
 }